financial-data-pipeline/requirements.txt
pandas
yfinance
SQLAlchemy
psycopg2-binary
python-dotenv
pytest
black
flake8
isort
